refactor: split HorizonSim787Definition into partial-class files
Move-only split of the 7,741-line HorizonSim787Definition.cs into a
partial class: BuildPanelControls -> .PanelControls.cs,
ProcessSimVarUpdate -> .SimVarUpdate.cs, HandleUIVariableSet +
HandleHotkeyAction -> .UiAndHotkeys.cs, and the aux-window/dialog
helpers -> .Dialogs.cs, per docs/design/god-file-split-map.md section 4.
No method bodies were altered; root keeps all fields and the
GetVariables()/GetPanelStructure()/TryGetDisplayOverride() cluster.
